Drill down into conditions for flags
This changes the Wagtail Flags UI to be a multi-level way to browse flags and their conditions. It allows you to edit existing conditions, delete conditions, and add new ones to each flag, as well as to add a flag itself in the UI.

It also hides "boolean" conditions behind an "Enable"/"Disable" set of buttons so that their purpose is clearer.