Make database-only flags deletable in 2.10+
This change makes it possible to delete database-only flags from within the admin in Wagtail 2.10 and above.

Feature flags should be ephemeral, and prior to this change it was not possible to delete database flags added with the "Add flag" button in the Wagtail admin without using the Django admin.

This change adds a button to the top of flag index page to delete the flag if it only exists as conditions in the database.