Add signals for when flag_enabled, flag_disabled are pressed
This change adds `flag_enabled` and `flag_disabled` signals. When an admin user presses the "Enable" or "Disable" button for a flag in the admin, the appropriate signal will be sent, enabling things like cache invalidation to happen based on the flag being enabled for all requests or disabled for all requests.

This is not an attempt to add general-purpose flag state signaling. That's more complex because state can be dependent on a request. But Wagtail-Flags provides global "Enable"/"Disable" buttons for flags, and this should allow signal handlers based on that.

For a little more context, please see #40.