Add backwards-compatible SVG breadcrumb icons
Wagtail 2.11 adds SVG icons to admin breadcrumbs. This change adds those SVG icons to the Flags admin pages.

It also adds a context variable to determine whether or not the SVG breadcrumb icons are available, and if not, falls back on the old style of breadcrumbs.

I've also taken the opportunity to define this context variable and 2.10's "action" item change in a single location in `views.py`.