Add interactive tox environment
This commit adds a new `tox -e interactive` that runs a local test server, to make it easier to test this project.

This is a feature that [wagtail-localize](https://www.wagtail-localize.org/#how-to-run-tests) has that we’ve also added to [wagtail-inventory](https://github.com/cfpb/wagtail-inventory#testing) and [wagtail-treemodeladmin](https://github.com/cfpb/wagtail-treemodeladmin). 🎩 to @chosak for this pattern.

Co-authored-by: Andy Chosak <andy.chosak@cfpb.gov>